move mutation queries to their own file
