+++ client/src/gql/queries.js
@@ -36,18 +36,6 @@ const GET_NOTE = gql`
 	}
 `;
 
-const SIGN_IN_USER = gql`
-	mutation signIn($email: String, $password: String!) {
-		signIn(email: $email, password: $password)
-	}
-`;
-
-const SIGNUP_USER = gql`
-	mutation signUp($email: String!, $username: String!, $password: String!) {
-		signUp(email: $email, username: $username, password: $password)
-	}
-`;
-
 const IS_LOGGED_IN = gql`
 	query WriteStatus {
 		status {
@@ -56,30 +44,4 @@ const IS_LOGGED_IN = gql`
 	}
 `;
 
-const NEW_NOTE = gql`
-	mutation newNote($content: String!) {
-		newNote(content: $content) {
-			id
-			content
-			createdAt
-			favouriteCount
-			favouriteBy {
-				id
-				username
-			}
-			author {
-				id
-				username
-			}
-		}
-	}
-`;
-
-export {
-	GET_NOTES,
-	GET_NOTE,
-	SIGN_IN_USER,
-	SIGNUP_USER,
-	IS_LOGGED_IN,
-	NEW_NOTE,
-};
+export { GET_NOTES, GET_NOTE, IS_LOGGED_IN };
